Introduce simple mechanism to blacklist resources by prefix
